Selenium tests
Hey everyone,

I have just made a commit to our GitHub project. I have added tests using Selenium which I believe is a better option as it simulates the user experience more realistically. In the code, I have used Selenium to automate the testing of the login and send message features on the app.

I have also made sure that the tests run in a headless mode to avoid any unnecessary overhead. Additionally, I have created a separate thread to start the server and waited for it to start up before the tests are executed.